requirements.txt
Django==1.5.2
South==0.8.2
distribute==0.6.24
dj-database-url==0.2.2
dj-static==0.0.5
django-extensions==1.2.0
django-social-auth==0.7.26
django-storages==1.1.8
django-tastypie==0.10.0
django-toolbelt==0.0.1
gunicorn==18.0
httplib2==0.8
mimeparse==0.1.3
newrelic==1.13.1.31
oauth2==1.5.211
psycopg2==2.5.1
python-dateutil==2.1
python-openid==2.2.5
simplejson==3.3.0
six==1.4.0
static==0.4
wsgiref==0.1.2
